Transform module inputs to tf json file (#29)
This creates the initial implementation of taking `PropertyMap` inputs
and converting it to a terraform json file (`*.tf.json`).

We could either create a `json` file or an `hcl` file. I decided to go
with the `json` format because it looks a lot simpler to convert the
inputs. Terraform module inputs seem to map pretty much directly to a
JSON representation.

See https://developer.hashicorp.com/terraform/language/syntax/json

closes #22