Run plan and parse results (#52)
This takes a first pass at extracting data from the Terraform plan in order to be able to call RegisterResource for each child resource in the module.

Currently this does not take into account any diff information in the plan and simply registers resource inputs based on the final plan value. In a future PR we will need to parse additional information from the Plan in order to compute an accurate diff for each resource.

re #9