Plan correctly handles unknown values
This updates the plan to correctly handle displaying unknown values.
Previously we were only determining properties based on the
`PlannedValues`, but for a plan that is incomplete. For example:

- On a `Create` it sometimes  does not contain properties that contain
  `unknown` values. This leads to properties being omitted from the
  plan.
- On an `Update` it will show the previous value from state if the
  property has become unknown

In order to correctly determine which properties are unknown we have to
look at the `ResourceChanges` and see if the property is in the
`AfterUnknown` list.

As part of this PR I also did a little refactoring of `details.go`.
Since we needed to do some more preprocessing on the plan results I
refactored some of the types to operate over a map of preprocessed
resources. This should also make resource lookup a little more efficient
later on. This will also be needed in my follow up PR when I process the
secret status of properties (which will affect both plan and state).