internal: use a plan store for the plan and state side channel (#104)
These changes harden the prototype implementation to be able to have 2 or more instances of a module in the same program. The in-memory side-channels used to communicate TF plans and states across the program are rewritten to be indexed on the URN of the top-most ModuleComponentResource. Names of the moduleStateResource and child resources are made less ambiguous to avoid "duplicate URN" errors.