Handle Pulumi secret inputs to the module
This PR adds support for first class Pulumi secrets as inputs to the
Terraform module. If a user creates a secret input (e.g.
`pulumi.secret(<value>)`) we will convert that to a `sensitive()` function in Terraform

There is some special encoding done for the value passed to the
`sensitive()` function based on what Terraform expects.

see
https://developer.hashicorp.com/terraform/language/syntax/json
https://developer.hashicorp.com/terraform/language/expressions/strings#string-templates

Once a input property is marked as sensitive, Terraform treats it as if
the variable is a sensitive type and it is tracked through to the child
resource inputs as well (which means we'll pick it back up from the
plan/state data).

closes #103