Convert Terraform `any` type to `pulumi.json#/Any` type (#131)
Module input variables that use the `any` type are currently being
converted to a `string` schema type. This PR changes this to convert the
`any` TF type to a `pulumi.json#/Any` type. I don't think it is possible
to know which specific type it is if a module uses the `any` type so I
think we have to use the Pulumi any type.

An example of this is the
[server_side_encryption_configuration](https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/blob/fc09cc6fb779b262ce1bee5334e85808a107d8a3/variables.tf#L211-L215)
property on the `s3_bucket` module. It has a `type = any` so it needs to
have a Pulumi any type.