Use existing tofu binary if it exists (#133)
This PR updates our acquisition of the tofu binary in several ways.

    We now first look to see if there is already a tofu binary installed
        by default it will check our install location as well as the system PATHs
    When we install tofu we now:
        Install to a common location in ~/.pulumi (similar to plugin installation)
        Install using lock files

I attempted to reuse the Pulumi plugin installation code from core, but it didn't completely work for our use case. I copied and modified some of the code from core that dealt with the final step of installing a file to a specific location on disk.

I've linked a todo to #71. This PR sets us up to handle further configuration on where to find the tofu binary and what version to look for. We could in the future do things like see which version of tofu a module requires and make sure one is installed or install the correct version otherwise.

For our tests I've added a GitHub actions step to install the tofu binary so it is available and each test won't try to re-download it.

closes #111