Mark module state as secret
The raw terraform state that we store in the ModuleState resource can
contain Terraform `sensitive` values which we convert to Pulumi `Secret`
values.

This PR takes the approach of marking the entire rawState `state`
property as secret. We proactively mark the entire state as secret,
regardless of whether it actually contains secret values. We could have
some additional processing to only do it if there is at least one
sensitive value in the state, but I don't think we would gain much by
doing this, and I think it is better to be consistent with how we store
the rawState in the Pulumi state.

closes #45