Fix number encoding from tofu plan and state json (#175)
The `terraform-json` library recommends setting `useJSONNumber(true)` to
avoid issues with precision loss. This PR enables this setting for
both `plan` and `apply` so that all numbers will be encoded as
`json.Number`.

This PR also adds some custom handling for `json.Number` when we convert
the tofu plan/state to `PropertyValue`s. We will convert to a `number`
if precision will be preserved, otherwise we will convert to a `string`.
This hopefully gives us the best of both worlds.

closes #168