Refresh support (#176)
This change implements support for pulumi refresh. Briefly, Pulumi calls Read method on the ModuleState resource, and the provider uses this trigger to run tofu refresh. When that is done the provider saves tofu plan and updated state, and then responds to Read(), Check() and Diff() calls from Pulumi for ModuleState and each child resource. If it is detected that the child resource is no longer there, Read() responds with a not-found marker and thereafter the resource is removed from state.

More background information on the workings on refresh is pending an edit to the Pulumi developer docs:
https://github.com/pulumi/pulumi/pull/18803