Use a shared cache directory for provider plugins (#185)
This configures a shared
[TF_PLUGIN_CACHE_DIR](https://opentofu.org/docs/cli/config/environment-variables/#tf_plugin_cache_dir)
for provider plugins. I've reused the same location that the dynamic
providers use to install terraform providers.

Tofu will now use that as the download location and will link it to the
local `.terraform` directory.

**Alternative Considered**

One alternative I considered was to instead use a shared [TF_DATA_DIR](https://opentofu.org/docs/cli/config/environment-variables/#tf_data_dir)
which would move the local `.terraform` directory to a shared location.
This would also mean we would cache the downloaded modules in this
shared location as well. I chose to go with the `CACHE_DIR` because it
was the more established pattern, but we could explore that more if we
want.

re #159