Fix inconsistent plan error
This is a temprorary(?)/partial(?) fix for the inconsistent plan error.

As part of creating the examples for (#203) I was running into this
issue and noticed that all of the occurrences were due to the
`terraform_data` resources changing from `null` to `""`. Adding this
simple `try` function to always fallback to `""` seems to have fixed the
issue in a lot of cases.

This obviously doesn't fix the inconsistent plan issue in all cases, but
it gets us unstuck for creating example tests.

re #198