Fix inconsistent plan error (#218)
This is a fix for the inconsistent plan error.

### Background

To expose outputs from a module, we need to account for the secretness of the outputs
If you try and output a secret value without setting `sensitive: true` in the output
The deployment will fail. We also need to be able to handle this dynamically since we won't know
the secret status until after the operation, and we need to write the output definition when we create the file (before the operation).

### Previous Solution

In #132 we added support for module outputs by using a `terraform_data` resource as a proxy for the output. This allowed us to account for the dynamic nature of the secret since we could determine the secret status of the value using the same method that we use for all other resource properties.

This ran into an issue #198 where `terraform_data` resources cannot be used if the value may be "inconsistent". A very common case that is used in most modules is to define an output like this:

```hcl
output "default_security_group_id" {
  description = "The ID of the security group created by default on VPC creation"
  value       = try(aws_vpc.this[0].default_security_group_id, null)
}
```

`try` is used because most modules will conditionally create resources and need to handle the case where the resource was not created and the value doesn't exist. This is a common case, but there may be other cases of this "inconsistent" data.

### New Solution

This PR tries a new solution of using output values with a combination of the [nonsensitive](https://developer.hashicorp.com/terraform/language/functions/nonsensitive) and [issensitive](https://developer.hashicorp.com/terraform/language/functions/issensitive) functions. We can use the `issensitive` function to dynamically determine whether the value is sensitive or not. Unfortunately it is not possible to use functions in the `output.sensitive` field (e.g. `sensitive: "${issensitive(module.source_module.output_name1)}"` won't work).

To work around this we use two outputs for each output value:
- The first output is the actual value that we mark as non-sensitive to avoid the error
- The second output is a boolean that indicates whether the value is a secret or not.

```json
"output": {
  "name1": {
      "value": "${nonsensitive(module.source_module.output_name1)}"
  },
  "internal_output_is_secret_name1": {
      "value": "${issensitive(module.source_module.output_name1)}"
  },
  ...
}
```

fixes #198