Restore terraform error logging
PR #204 changed the way we handle logs and we lost error logs when
something fails in Terraform. To handle the error messages I've brought
in some `opentofu` code for handling logging. As a side effect this
could also allow us in the future to specifically handle certain types
of errors.

We still don't get the Terraform plan in the debug logs, but we can add
that later.

fixes #216