Remove TF_PLUGIN_CACHE_DIR usage (#223)
I keep running into CI flakes with an error on the `Test_TwoInstances_TypeScript` test.

```console
 error: randmod:index:Module resource 'myrandmod2' has a problem: NewModuleComponentResource failed: Init failed: error running tofu init: exit status 1
```

It seems to happen when there is not a already existing temp folder for both modules (which is always the case for CI). I think it is due to using the shared plugin cache so turning it off for now so I can revisit it after our CI stops failing.

re #199