Fix unknown propagation (#278)
This change fixes propagating Pulumi unknowns to Terraform module positions of complex type (such as list, object types). Before the change the solution was using a terraform_data resource to emit unknown literals into HCL sources. This worked for scalar types but not for compound types such as lists and objects. After the change, the solution is using a special purpose-built Terraform provider to inject unknown values of Dynamic type. The provider is hosted in the same process as the module provider and linked dynamically into the Tofu sandbox using ReattachConfig.

Fixes #28

Does not fix #270 - that remains a problem.