Unwrap secreted configs in cleanProvidersConfig (#291)
This pull request adds the ability to unwrap a secreted config value in cleanProvidersConfig.

Currently, a config object wrapped in a secret would fail as an unsupported type. But part of the logic in pulumi/pulumi-terraform-bridge#3007 wraps the terraformConfig().return object in a secret. This object is already properly formatted to be usable by Terraform; so all we need to do is return the secret's value.

A potential concern is that any user-provided secret config inputs would also hit this line, and perhaps not get cleaned up properly. I'm not sure how much of a concern that should be - the alternative is running two separate passes over s.providerConfig.

Prerequisite for pulumi/pulumi-terraform-bridge#3007 .