chore: fix pulumi codegen pin (#321)
Investigating why #319 PR is failing we probably are not correctly pinning Pulumi CLI in the TestGenerateTerraformAwsModulesSDKs test. Somehow the Pulumi CLI version from ESC toolscache path is starting to interfere. This PR is an attempt to get this more robust so that .pulumi/bin/pulumi pinned version is used exactly.

The code churn is from upgrading the test data to newer Pulumi CLI versions.