Fix TestLambdaMemorySizeDiff
The diff displayed under Views is actually correct, memory_size is added as a module input but rendered as updated on
the view resource representing the Lambda Function, since currently all view properties are modeled as inputs it
detects a change from the provider-assigned value to the new user-assigned value.

Since there are still problems with Views in update plans, the test drops down to regexes to assert the result.