dependencies: pulumi v3.176.0 (#402)
* dependencies: pulumi v3.176.0

Update Pulumi to v3.176.0. With this version, the feature flag no longer needs to be set; views are enabled by default.

* Set `PULUMI_ENABLE_VIEWS_PREVIEW` for YAML

We need to release a new version of pulumi-yaml, pull that new version into pulumi/pulumi, and then release a new version of pulumi/pulumi to avoid the need to set the envvar for YAML programs.