Unskip Test_replace_drift_deleted (#401)
* t0yv0/pulumi-v3.175.0 (#373)

* Better missing credentials error when using AWS modules (#375)

* Better missing credentials error when using AWS modules

* embed error message and link to configuring explicit providers

* Update README for executor changes. (#388)

* non-nullable inputs should be marked as required (#385)

* non-nullable inputs should be marked as required

* add opentofu to the executors of the test

* Fix VPC module typings, include outputs that refer to serialized resources (#393)

* Default output type is any, better inference for try(...) expressions (#398)

* Unskip Test_replace_drift_deleted

The issue in the CLI has been addressed and released, so we can unskip this test.

* dependencies: pulumi v3.176.0 (#402)

* dependencies: pulumi v3.176.0

Update Pulumi to v3.176.0. With this version, the feature flag no longer needs to be set; views are enabled by default.

* Set `PULUMI_ENABLE_VIEWS_PREVIEW` for YAML

We need to release a new version of pulumi-yaml, pull that new version into pulumi/pulumi, and then release a new version of pulumi/pulumi to avoid the need to set the envvar for YAML programs.

---------

Co-authored-by: Anton Tayanovskyy <anton@pulumi.com>
Co-authored-by: Zaid Ajaj <Zaid-Ajaj@users.noreply.github.com>
Co-authored-by: VenelinMartinov <venelin.v.martinov@gmail.com>
Co-authored-by: Zaid Ajaj <zaid.naom@gmail.com>