Specify RefreshBeforeUpdate on partial failures
Pulumi v3.178.0 now supports indicating a resource should be refreshed by default from partial failures. Previously, it was only possible to indicate this on successful provider operations.

With this, if a Module CustomResource partially fails during Create (e.g. one of it's children has failed), it can report a partial failure using `ErrorResourceInitFailed` and now indicate that it should be refreshed before being updated. This way, on the next `pulumi up`, the Module resource will be refreshed before `Update` as intended, giving the Module resource an opportunity to "retry" for any of the failed children.