Specify RefreshBeforeUpdate on partial failures (#424)
This PR updates dependencies to pulumi v3.178.0 and specifies RefreshBeforeUpdate on partial failures (which is now supported in v3.178.0). Previously, it was only possible to indicate RefreshBeforeUpdate on successful provider operations.

With this, if a Module CustomResource partially fails during Create (e.g. one of it's children has failed), as part of reporting a partial failure using `ErrorResourceInitFailed`, it can now indicate that it should be refreshed by default. This way, on the next `pulumi up`, the Module resource will be refreshed as intended, giving the Module resource an opportunity to "retry" for any of the failed children.