Report partial errors on create and update (#428)
* Report partial errors

* test partial errors

* Fix apply error

Use a different name than `err` when calling `applyModuleOperation `, otherwise `err` will be overwritten by the call to `PublishViewSteps`.

* More logging

* Detect more error cases for views

This definitely could be refactored/improved (it's not really the right place to be eliding the error), but it does correctly handle the `ReplaceDestroyBeforeCreate` case when delete succeeds but create fails (we assume).

* clean up viewStepStatusCheck

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>