Run init with `-upgrade` when module version is updated
When a user upgrades a module, that module can require a newer version
of the underlying Terraform providers. When this happens you have to run
`tofu init -upgrade` in order to upgrade the version of the provider
that is used.

This PR updates the provider to save the module version to state so that
we can compare to the new version and `-update` if necessary.

fixes #547